requirements.txt
Django==2.0.7
djangorestframework==3.8.2
gunicorn==19.9.0
mysqlclient==1.3.12
Pillow==5.2.0
pytz==2018.5
